add rolse and permission

+++ api/handlers.py
@@ -7,15 +7,18 @@ from fastapi import HTTPException
 from sqlalchemy.exc import IntegrityError
 from sqlalchemy.ext.asyncio import AsyncSession
 
+from api.actions.auth import get_current_user_from_token
 from api.actions.user import _create_new_user
 from api.actions.user import _delete_user
 from api.actions.user import _get_user_by_id
 from api.actions.user import _update_user
+from api.actions.user import check_user_permissions
 from api.schemas import DeleteUserResponse
 from api.schemas import ShowUser
 from api.schemas import UpdateUserRequest
 from api.schemas import UpdateUserResponse
 from api.schemas import UserCreate
+from db.models import User
 from db.session import get_db
 
 
@@ -34,8 +37,20 @@ async def create_user(body: UserCreate, db: AsyncSession = Depends(get_db)) -> S
 
 @user_router.delete("/", response_model=DeleteUserResponse)
 async def delete_user(
-    user_id: UUID, db: AsyncSession = Depends(get_db)
+    user_id: UUID,
+    db: AsyncSession = Depends(get_db),
+    current_user: User = Depends(get_current_user_from_token)
 ) -> DeleteUserResponse:
+    user_for_deletion = await _get_user_by_id(user_id, db)
+    if user_for_deletion is None:
+        raise HTTPException(
+            status_code=404, detail=f"User with id {user_id} not found."
+        )
+    if not check_user_permissions(
+            target_user=user_for_deletion,
+            current_user=current_user,
+    ):
+        raise HTTPException(status_code=403, detail="Forbidden.")
     deleted_user_id = await _delete_user(user_id, db)
     if deleted_user_id is None:
         raise HTTPException(
@@ -45,7 +60,11 @@ async def delete_user(
 
 
 @user_router.get("/", response_model=ShowUser)
-async def get_user_by_id(user_id: UUID, db: AsyncSession = Depends(get_db)) -> ShowUser:
+async def get_user_by_id\
+                (user_id: UUID,
+                 db: AsyncSession = Depends(get_db),
+                 current_user: User = Depends(get_current_user_from_token),
+) -> ShowUser:
     user = await _get_user_by_id(user_id, db)
     if user is None:
         raise HTTPException(
@@ -56,7 +75,8 @@ async def get_user_by_id(user_id: UUID, db: AsyncSession = Depends(get_db)) -> S
 
 @user_router.patch("/", response_model=UpdateUserResponse)
 async def update_user_by_id(
-    user_id: UUID, body: UpdateUserRequest, db: AsyncSession = Depends(get_db)
+    user_id: UUID, body: UpdateUserRequest, db: AsyncSession = Depends(get_db),
+    current_user: User = Depends(get_current_user_from_token),
 ) -> UpdateUserResponse:
     updated_user_params = body.dict(exclude_none=True)
     if updated_user_params == {}:
@@ -64,11 +84,17 @@ async def update_user_by_id(
             status_code=422,
             detail="At least one parameter for user update info should be provided",
         )
-    user = await _get_user_by_id(user_id, db)
-    if user is None:
+    user_for_update = await _get_user_by_id(user_id, db)
+    if user_for_update is None:
         raise HTTPException(
             status_code=404, detail=f"User with id {user_id} not found."
         )
+    if user_id != current_user.user_id:
+        if check_user_permissions(
+            target_user=user_for_update,
+            current_user=current_user,
+        ):
+            raise HTTPException(status_code=403, detail="Forbidden")
     try:
         updated_user_id = await _update_user(
             updated_user_params=updated_user_params, session=db, user_id=user_id
